requirements.txt
django
djangorestframework
psycopg2-binary
faker
drf-spectacular